+++ fake-cucumber/javascript/src/IParameterTypeDefinition.ts
@@ -1,0 +1,8 @@
+export default interface IParameterTypeDefinition {
+  name: string
+  regexp: RegExp | RegExp[] | string | string[]
+  type?: any
+  transformer?: (...args: string[]) => any
+  preferForRegexpMatch?: boolean
+  useForSnippets?: boolean
+}