+++ fake-cucumber/javascript/src/SupportCode.ts
@@ -4,52 +4,18 @@ import {
   ParameterTypeRegistry,
 } from '@cucumber/cucumber-expressions'
 import { IdGenerator, messages } from '@cucumber/messages'
-import StackUtils from 'stack-utils'
 import { AnyBody } from './types'
 import ExpressionStepDefinition from './ExpressionStepDefinition'
 import IStepDefinition from './IStepDefinition'
 import IHook from './IHook'
 import Hook from './Hook'
 import IClock from './IClock'
-import { MakeErrorMessage } from './ErrorMessageGenerator'
-
-type RegisterStepDefinition = (
-  expression: string | RegExp,
-  body: AnyBody
-) => void
-
-type RegisterHook = (
-  tagExpressionOrBody: string | AnyBody,
-  body?: AnyBody
-) => void
-
-interface IParameterTypeDefinition {
-  name: string
-  regexp: RegExp | RegExp[] | string | string[]
-  type?: any
-  transformer?: (...args: string[]) => any
-  preferForRegexpMatch?: boolean
-  useForSnippets?: boolean
-}
+import { MakeErrorMessage, withFullStackTrace } from './ErrorMessageGenerator'
+import IParameterTypeDefinition from './IParameterTypeDefinition'
+import PerfHooksClock from './PerfHooksClock'
 
 function defaultTransformer(...args: string[]) {
   return args
-}
-
-function getSourceReference(stackTrace: string): messages.ISourceReference {
-  const stack = new StackUtils({
-    cwd: process.cwd(),
-    internals: StackUtils.nodeInternals(),
-  })
-  const trace = stack.clean(stackTrace)
-  const callSite = stack.parseLine(trace.split('\n')[1])
-  const { file: uri, line } = callSite
-  return new messages.SourceReference({
-    uri,
-    location: new messages.Location({
-      line,
-    }),
-  })
 }
 
 /**
@@ -60,17 +26,7 @@ export default class SupportCode {
   public readonly stepDefinitions: IStepDefinition[] = []
   public readonly beforeHooks: IHook[] = []
   public readonly afterHooks: IHook[] = []
-  public readonly Given = this.registerStepDefinition.bind(
-    this
-  ) as RegisterStepDefinition
-  public readonly When = this.registerStepDefinition.bind(
-    this
-  ) as RegisterStepDefinition
-  public readonly Then = this.registerStepDefinition.bind(
-    this
-  ) as RegisterStepDefinition
-  public readonly Before = this.registerBeforeHook.bind(this) as RegisterHook
-  public readonly After = this.registerAfterHook.bind(this) as RegisterHook
+
   private readonly parameterTypeRegistry = new ParameterTypeRegistry()
   private readonly expressionFactory = new ExpressionFactory(
     this.parameterTypeRegistry
@@ -78,9 +34,9 @@ export default class SupportCode {
   public readonly undefinedParameterTypes: messages.IEnvelope[] = []
 
   constructor(
-    public newId: IdGenerator.NewId,
-    public clock: IClock,
-    public makeErrorMessage: MakeErrorMessage
+    public readonly newId: IdGenerator.NewId = IdGenerator.uuid(),
+    public readonly clock: IClock = new PerfHooksClock(),
+    public readonly makeErrorMessage: MakeErrorMessage = withFullStackTrace()
   ) {}
 
   public defineParameterType(
@@ -98,11 +54,11 @@ export default class SupportCode {
     this.parameterTypes.push(parameterType)
   }
 
-  private registerStepDefinition(
+  public defineStepDefinition(
+    sourceReference: messages.ISourceReference,
     expression: string | RegExp,
     body: AnyBody
   ): void {
-    const sourceReference = getSourceReference(new Error().stack)
     try {
       const expr = this.expressionFactory.createExpression(expression)
       const stepDefinition = new ExpressionStepDefinition(
@@ -128,30 +84,34 @@ export default class SupportCode {
     }
   }
 
-  private registerBeforeHook(
+  public defineBeforeHook(
+    sourceReference: messages.ISourceReference,
     tagExpressionOrBody: string | AnyBody,
     body?: AnyBody
   ) {
-    this.beforeHooks.push(this.makeHook(new Error(), tagExpressionOrBody, body))
+    this.beforeHooks.push(
+      this.makeHook(sourceReference, tagExpressionOrBody, body)
+    )
   }
 
-  private registerAfterHook(
+  public defineAfterHook(
+    sourceReference: messages.ISourceReference,
     tagExpressionOrBody: string | AnyBody,
     body?: AnyBody
   ) {
-    this.afterHooks.push(this.makeHook(new Error(), tagExpressionOrBody, body))
+    this.afterHooks.push(
+      this.makeHook(sourceReference, tagExpressionOrBody, body)
+    )
   }
 
   private makeHook(
-    error: Error,
+    sourceReference: messages.ISourceReference,
     tagExpressionOrBody: string | AnyBody,
     body?: AnyBody
   ) {
     const tagExpression =
       typeof tagExpressionOrBody === 'string' ? tagExpressionOrBody : null
     body = typeof tagExpressionOrBody !== 'string' ? tagExpressionOrBody : body
-
-    const sourceReference = getSourceReference(error.stack)
     return new Hook(this.newId(), tagExpression, sourceReference, body)
   }
 }