+++ gherkin/javascript/src/Gherkin.ts
@@ -15,7 +15,11 @@ function fromStream(stream: Readable, options: IGherkinOptions) {
   )
 }
 
-function fromPaths(paths: string[], options: IGherkinOptions): Readable {
+function fromPaths(
+  paths: ReadonlyArray<string>,
+  options: IGherkinOptions
+): Readable {
+  const pathsCopy = paths.slice()
   options = makeGherkinOptions(options)
   const combinedMessageStream = new PassThrough({
     writableObjectMode: true,
@@ -23,14 +27,14 @@ function fromPaths(paths: string[], options: IGherkinOptions): Readable {
   })
 
   function pipeSequentially() {
-    const path = paths.shift()
+    const path = pathsCopy.shift()
     if (path !== undefined) {
       const parserMessageStream = new ParserMessageStream(options)
       parserMessageStream.on('end', () => {
         pipeSequentially()
       })
 
-      const end = paths.length === 0
+      const end = pathsCopy.length === 0
       // Can't use pipeline here because of the { end } argument,
       // so we have to manually propagate errors.
       options
@@ -48,9 +52,10 @@ function fromPaths(paths: string[], options: IGherkinOptions): Readable {
 }
 
 function fromSources(
-  envelopes: messages.IEnvelope[],
+  envelopes: ReadonlyArray<messages.IEnvelope>,
   options: IGherkinOptions
 ): Readable {
+  const envelopesCopy = envelopes.slice()
   options = makeGherkinOptions(options)
   const combinedMessageStream = new PassThrough({
     writableObjectMode: true,
@@ -58,11 +63,11 @@ function fromSources(
   })
 
   function pipeSequentially() {
-    const envelope = envelopes.shift()
+    const envelope = envelopesCopy.shift()
     if (envelope !== undefined && envelope.source) {
       const parserMessageStream = new ParserMessageStream(options)
       parserMessageStream.pipe(combinedMessageStream, {
-        end: envelopes.length === 0,
+        end: envelopesCopy.length === 0,
       })
       parserMessageStream.on('end', pipeSequentially)
       parserMessageStream.end(envelope)